% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Computes the Akaike information criterion of an ordinary least square regression model.
The AIC was corrected for synthetically oversampled data, for which the sample size is unrepresentable}
\usage{
aic(model, oversampled = FALSE, amount = 0)
}
\arguments{
\item{model}{An ordinary least square regression model}

\item{oversampled}{A logical value indicating whether the data set used to train the regression model was oversampled or not.}

\item{amount}{The number of synthetically generated observations added to the oversampled data set}
}
\value{
A numeric AIC value calculated as $n log(SSE) - n log(n) + 2p$
}
\description{
Computes the Akaike information criterion of an ordinary least square regression model.
The AIC was corrected for synthetically oversampled data, for which the sample size is unrepresentable
}
