% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotMissing.R
\name{boxplotMissing}
\alias{boxplotMissing}
\title{This function was written to impute and explore missing data. It uses Multivariate Imputation Chained Equations to perform this.}
\usage{
boxplotMissing(df, makePlot = TRUE)
}
\arguments{
\item{df}{A data frame as given by the owner of the data}

\item{makePlot}{logical, if TRUE, a plot will be generated and stored in file "boxplotMissing.pdf"}
}
\value{
A similar data frame with MICE imputed missing values for met_002, met_010 and met_068.
}
\description{
This function was written to impute and explore missing data. It uses Multivariate Imputation Chained Equations to perform this.
}
