% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFraction.R
\name{calculateFraction}
\alias{calculateFraction}
\title{Estimate the fraction of category 2 + 95pc CI from a sample with two-category outcome counts.
Return result in character format}
\usage{
calculateFraction(n1, n2, reps = 1000)
}
\arguments{
\item{n1}{Number of observed category 1 outcomes}

\item{n2}{Number of observed category 2 outcomes}

\item{reps}{Number of bootstrap replicates used to estimate the 95pc CI, defaults to 1000}
}
\value{
An estimate of the fraction + 95pc CI in character format
}
\description{
Estimate the fraction of category 2 + 95pc CI from a sample with two-category outcome counts.
Return result in character format
}
\examples{
calculateFraction(n1=90, n2=10)

}
