% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFractionNumeric.R
\name{calculateFractionNumeric}
\alias{calculateFractionNumeric}
\title{Estimate the fraction of category 2 + 95pc CI from a sample with two-category outcome counts.
Return result in numeric format}
\usage{
calculateFractionNumeric(n1, n2, reps = 1000)
}
\arguments{
\item{n1}{Number of observed category 1 outcomes}

\item{n2}{Number of observed category 2 outcomes}

\item{reps}{Number of bootstrap replicates used to estimate the 95pc CI, defaults to 1000}
}
\value{
An estimate of the fraction + 95pc CI as a numeric vector
}
\description{
Estimate the fraction of category 2 + 95pc CI from a sample with two-category outcome counts.
Return result in numeric format
}
\examples{
calculateFractionNumeric(n1=90, n2=10)

}
