% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntMedian.R
\name{confIntMedian}
\alias{confIntMedian}
\title{Takes a wilcox.test object with option "conf.int=TRUE" and returns the pseudo-median
difference between two population in character format.}
\usage{
confIntMedian(wilcoxTest, negative = FALSE)
}
\arguments{
\item{wilcoxTest}{a wilcox.test object with option "conf.int=TRUE"}

\item{negative}{Logical adressing whether the roles of the two groups must be reversed}
}
\value{
An estimate of the pseudo median difference + 95pc CI
}
\description{
Takes a wilcox.test object with option "conf.int=TRUE" and returns the pseudo-median
difference between two population in character format.
}
\examples{

dataA <- data.frame(group=rep("A", times=20),
                    val=rnorm(n=20, mean=0, sd=1))
dataB <- data.frame(group=rep("B", times=20),
                    val=rnorm(n=20, mean=1, sd=1))
data <- bind_rows(dataA, dataB)
wilcoxTest <- wilcox.test(formula=val~group, data=data, conf.int=TRUE)
confIntMedian(wilcoxTest, negative=FALSE)
confIntMedian(wilcoxTest, negative=TRUE)

}
