% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToTexTable.R
\name{convertToTexTable}
\alias{convertToTexTable}
\title{Helper function that writes results to a .tex file}
\usage{
convertToTexTable(
  tab,
  filename,
  caption = NULL,
  reflabel = NULL,
  rows.named = FALSE,
  minipage = FALSE
)
}
\arguments{
\item{tab}{matrix or data frame}

\item{filename}{Name of the file the matrix or data frame must be written to. Using a .tex extension is advised, suitability for other file extensions was not checked}

\item{caption}{Information and explanations about the table as a character string.}

\item{reflabel}{Label of the .tex table for referencing as a character string.}

\item{rows.named}{Logical, set to TRUE for matrices only. If TRUE, The rownames of the matrix are included in the .tex file}

\item{minipage}{Logical, set to TRUE if the .tex table is included in a "minipage" environment in LaTeX}
}
\description{
Write a matrix or data frame to a .tex table, ready to include in your LaTeX report. Specify different options
dependent on the context of the file in the LaTeX report.
}
\examples{

getwd()

tab <- matrix(1:9, ncol=3)
colnames(tab) <- c("col1", "col2", "col3")
rownames(tab) <- c("row1", "row2", "row3")

convertToTexTable(tab, "tab.tex")
convertToTexTable(tab, "tabWithCaption.tex", caption="This is a test", reflabel="test_matrix")
convertToTexTable(tab, "tabWithRowsNamed.tex", caption="This is a test", reflabel="test_matrix", rows.named=TRUE)
convertToTexTable(tab, "tabMinipage.tex", minipage=TRUE)

}
