% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countClassesNumeric.R
\name{countClassesNumeric}
\alias{countClassesNumeric}
\title{Summarizes a patients and their reclassification status to counts and fractions reclassified as syndromatic.
The point estimate and the confidence intervals are passed as three numeric columns}
\usage{
countClassesNumeric(data, race, strat)
}
\arguments{
\item{data}{A data frame containing columns "ObesityClass" and "Reclassified", the latter consisting of values "Healthy" and "Syndromatic"}

\item{race}{Ethnicity of the patients in the data set}

\item{strat}{Stratification of the model used for reclassification}
}
\value{
A tibble with columns "Race", "model", "ObesityClass", "pred.: Healthy", "pred. Syndr." and "frac. syndromatic"
}
\description{
Summarizes a patients and their reclassification status to counts and fractions reclassified as syndromatic.
The point estimate and the confidence intervals are passed as three numeric columns
}
\examples{
normals <- data.frame(ObesityClass=rep("Normal weight", times=50),
                      Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.2)])
overweights <- data.frame(ObesityClass=rep("Overweight", times=50),
                          Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.5)])
obeses <- data.frame(ObesityClass=rep("Obese", times=50),
                          Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.8)])
data <- bind_rows(normals, overweights, obeses)
countClassesNumeric(data, race="Not specified", strat="Unknown")

}
