% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonstrateOutlierDetection.R
\name{demonstrateOutlierDetection}
\alias{demonstrateOutlierDetection}
\title{Stores a plot to demonstrate outlier detection with metabolites and ethnicities of interest, stored in the file "outlierDemonstration.pdf"}
\usage{
demonstrateOutlierDetection(df, mets, ethnicity)
}
\arguments{
\item{df}{A data frame containing columns "ID", "BMI", "Race" and metabolite levels.}

\item{mets}{Metabolite names of interest}

\item{ethnicity}{Ethnicities according to the metabolites of interest}
}
\description{
Stores a plot to demonstrate outlier detection with metabolites and ethnicities of interest, stored in the file "outlierDemonstration.pdf"
}
