% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrNO.R
\name{illustrNO}
\alias{illustrNO}
\title{Illustrates the interpretation of the figures on metabolic differences of the NO outliers}
\usage{
illustrNO()
}
\value{
A ggplot object
}
\description{
Illustrates the interpretation of the figures on metabolic differences of the NO outliers
}
