% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrON.R
\name{illustrON}
\alias{illustrON}
\title{Illustrates the interpretation of the figures on metabolic differences of the ON outliers}
\usage{
illustrON()
}
\value{
A ggplot object
}
\description{
Illustrates the interpretation of the figures on metabolic differences of the ON outliers
}
