% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormalityAndOutliers.R
\name{logNormalityAndOutliers}
\alias{logNormalityAndOutliers}
\title{Assess log-normality of metabolite distributions + tagging outliers}
\usage{
logNormalityAndOutliers(df)
}
\arguments{
\item{df}{A data frame containing columns "ID", "Race", "BMI" and metabolite levels.}
}
\value{
A list object containing a data frame with outliers for every metabolite, a count of outliers + frequency of common outliers with other metabolites, a table of metabolites which were not log-normally distributed + p-value of the Shapiro-Wilk test
}
\description{
Assess log-normality of metabolite distributions + tagging outliers
}
