% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQQplot.R
\name{makeQQplot}
\alias{makeQQplot}
\title{Makes QQ-plots of a subselection of metabolites, stored in the file "QQplots.pdf"}
\usage{
makeQQplot(df, mets, pvals)
}
\arguments{
\item{df}{A data frame containing columns "ID", "BMI", "Race" and metabolite levels.}

\item{mets}{A subselection of metabolites to make a QQ-plot of}
}
\description{
Makes QQ-plots of a subselection of metabolites, stored in the file "QQplots.pdf"
}
