% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDiagnostics.R
\name{modelDiagnostics}
\alias{modelDiagnostics}
\title{Stores a pdf file called paste0("modelDiagnostics", ethnicity, ".pdf") with plots of the residuals and squared residuals as a function of the predicted BMI.
The functional form of the predictors significantly deviating from linearity are included.
All p-values are calculated from F-tests between an loess-estimator and an intercept estimator of the residuals}
\usage{
modelDiagnostics(
  effects,
  types,
  transformations,
  balancing = NULL,
  ethnicity,
  model = NULL,
  new.data
)
}
\arguments{
\item{effects}{a vector of "main" and/or "interaction" effects of formulations the model diagnostics need to be examined of}

\item{types}{a vector of regression types "OLS", "Ridge" and/or "LASSO" of formulations the model diagnostics need to be examined of}

\item{balancing}{defaults to NULL, or a vector of "" and/or "Balanced" reflecting balancing requests of formulations the model diagnostics need to be examined of}

\item{ethnicity}{the ethnicity of patients in new.data}

\item{new.data}{Data to build a regression model from}

\item{transformation}{a vector of BMI transformations "Log" and/or "Inv" of formulations the model diagnostics need to be examined of}
}
\description{
Stores a pdf file called paste0("modelDiagnostics", ethnicity, ".pdf") with plots of the residuals and squared residuals as a function of the predicted BMI.
The functional form of the predictors significantly deviating from linearity are included.
All p-values are calculated from F-tests between an loess-estimator and an intercept estimator of the residuals
}
