% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample.R
\name{oversample}
\alias{oversample}
\title{Oversamples a given data set to balance out the "Overweight" and "Obese" obesity classes with respect to the "Normal weight" using SMOTE.
The data set must contain the two columns called "ObesityClass" and "met_110",  and some columns of the globally specified "metabolites", "BMI" and/or "Age".}
\usage{
oversample(data_train)
}
\arguments{
\item{data_train}{A data frame. One of the colums must be called "ObesityClass", consisting of values "Normal weight", "Overweight" and "Obese". Some columns must be}
}
\value{
A SMOTE oversampled version of the data set as a data frame. Patient ID and other irrelevant columns are not included
}
\description{
Oversamples a given data set to balance out the "Overweight" and "Obese" obesity classes with respect to the "Normal weight" using SMOTE.
The data set must contain the two columns called "ObesityClass" and "met_110",  and some columns of the globally specified "metabolites", "BMI" and/or "Age".
}
\examples{
metabolites <- c("met_001", "met_002", "met_110", "met_coll")

oversampled_data <- oversample(df)

table(oversampled_data$ObesityClass)
boxplot(BMI~ObesityClass, oversampled_data)
boxplot(BMI~ObesityClass, df)

}
