% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotANOVA.R
\name{plotANOVA}
\alias{plotANOVA}
\title{Calculates the difference in metabolic levels between ON, NO, OO and NN.
A plot is made of the scaled metabolic differences for comparison.}
\usage{
plotANOVA(data, ethnicity)
}
\arguments{
\item{data}{Data with metabolite levels and assigned prediction group "predictionGroup"}

\item{ethnicity}{Ethnicity of the patients in the data set}
}
\value{
A list of 2 ggplot objects for NO ("normal") and ON ("obese") outliers
}
\description{
Calculates the difference in metabolic levels between ON, NO, OO and NN.
A plot is made of the scaled metabolic differences for comparison.
}
\examples{

metabolites <- c("met1", "met2", "met_coll", "met_110")

model <- trainRidgeLASSO(effect="main", type="Ridge", transformation="Log", new.data=df)
df$predicted <-
  exp(predict(model, newx=makeMatrix(df, includeInteraction=FALSE)$mat)[,"s0"])

rocTrain <- roc(df$ObesityClass, df$predicted,
                levels=c("Normal weight", "Obese"))
cutoff <- coords(rocTrain, x="best")[1,"threshold"]
df <- df \%>\%
  mutate(predictionGroup = ifelse(ObesityClass=="Normal weight"&predicted<cutoff, "NN",
                             ifelse(ObesityClass=="Normal weight"&predicted>cutoff, "NO",
                               ifelse(ObesityClass=="Obese"&predicted>cutoff, "OO",
                                 ifelse(ObesityClass=="Obese"&predicted<cutoff, "ON", NA)))))

ggplot(df, aes(x=predicted, y=exp(BMI), col=predictionGroup)) +
  geom_point() +
  labs(x="predicted BMI", y="observed BMI")

plotANOVA(df, "White")

}
