% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScaledEffects.R
\name{plotScaledEffects}
\alias{plotScaledEffects}
\title{Visualize the regression coefficients + 95 pc CI of the bootstrap replicates.
The groups of correlated metabolites were specified from the inspection of hospital data and are not customizable in this version of the function.}
\usage{
plotScaledEffects(
  bootstrapCoeffsWhite,
  bootstrapCoeffsBlack,
  effect,
  type,
  transformation
)
}
\arguments{
\item{bootstrapCoeffsWhite}{Bootstrap replicated coefficients of the white patients data}

\item{bootstrapCoeffsBlack}{Bootstrap replicated coefficients of the black patients data}

\item{effect}{"main" or "interaction" dependent on which effects are wished}

\item{type}{Regression type, must be "Ridge" or "LASSO for this function}

\item{transformation}{"Log" or "Inv", dependent on whether regression on log(BMI) or 1/BMI is required}
}
\value{
If model only uses main effects, a ggplot2 plot is returned of the main effects, if interaction effects used, a list of two plots for main effects and 50 most important interaction effects.
}
\description{
Visualize the regression coefficients + 95 pc CI of the bootstrap replicates.
The groups of correlated metabolites were specified from the inspection of hospital data and are not customizable in this version of the function.
}
\examples{
metabolites <- c("met1", "met2", "met_coll", "met_110")

data_normal_white <- data.frame(ObesityClass=rep("Normal weight", times=200),
                          Age=rnorm(n=200, mean=30, sd=3),
                          BMI=runif(n=200, min=18, max=25),
                          met1=rnorm(n=200, mean=0.5, sd=0.5),
                          met2=rnorm(n=200, mean=5, sd=1),
                          met_110=rnorm(n=200, mean=100, sd=100))
data_overweight_white <- data.frame(ObesityClass=rep("Overweight", times=100),
                          Age=rnorm(n=100, mean=30, sd=3),
                          BMI=runif(n=100, min=25, max=30),
                          met1=rnorm(n=10, mean=0.7, sd=0.5),
                          met2=rnorm(n=100, mean=4, sd=1),
                          met_110=rnorm(n=100, mean=100, sd=100))
data_obese_white <- data.frame(ObesityClass=rep("Obese", times=50),
                          Age=rnorm(n=50, mean=30, sd=3),
                          BMI=runif(n=50, min=30, max=40),
                          met1=rnorm(n=50, mean=1, sd=0.5),
                          met2=rnorm(n=50, mean=3, sd=1),
                          met_110=rnorm(n=50, mean=100, sd=100))

data_train_white <- bind_rows(data_normal_white, data_overweight_white, data_obese_white)
data_train_white$BMI <- log(data_train_white$BMI)
data_train_white$met_coll <- data_train_white$met1 + 7.2*data_train_white$met2 + rnorm(n=nrow(data_train_white), mean=0, sd=0.1)

data_normal_black <- data.frame(ObesityClass=rep("Normal weight", times=200),
                          Age=rnorm(n=200, mean=30, sd=3),
                          BMI=runif(n=200, min=18, max=25),
                          met1=rnorm(n=200, mean=0.5, sd=0.5),
                          met2=rnorm(n=200, mean=7, sd=1),
                          met_110=rnorm(n=200, mean=100, sd=100))
data_overweight_black <- data.frame(ObesityClass=rep("Overweight", times=100),
                          Age=rnorm(n=100, mean=30, sd=3),
                          BMI=runif(n=100, min=25, max=30),
                          met1=rnorm(n=10, mean=0.7, sd=0.5),
                          met2=rnorm(n=100, mean=5, sd=1),
                          met_110=rnorm(n=100, mean=100, sd=100))
data_obese_black <- data.frame(ObesityClass=rep("Obese", times=50),
                          Age=rnorm(n=50, mean=30, sd=3),
                          BMI=runif(n=50, min=30, max=40),
                          met1=rnorm(n=50, mean=1, sd=0.5),
                          met2=rnorm(n=50, mean=3, sd=1),
                          met_110=rnorm(n=50, mean=100, sd=100))

data_train_black <- bind_rows(data_normal_black, data_overweight_black, data_obese_black)
data_train_black$BMI <- log(data_train_black$BMI)
data_train_black$met_coll <- data_train_black$met1 + 7.2*data_train_black$met2 + rnorm(n=nrow(data_train_black), mean=0, sd=0.1)


bootstrapCoeffsWhite <- scaledEffects(data_train_white, effect="main",
                                      type="LASSO", transformation="Log",
                                      balancing="Balanced", boot.n=1000)
bootstrapCoeffsWhite <- pivot_longer(bootstrapCoeffsWhite,
                                     cols=colnames(bootstrapCoeffsWhite),
                                     names_to="met", values_to="coeff")
bootstrapCoeffsBlack <- scaledEffects(data_train_black, effect="main",
                                      type="LASSO", transformation="Log",
                                      balancing="Balanced", boot.n=1000)
bootstrapCoeffsBlack <- pivot_longer(bootstrapCoeffsBlack,
                                     cols=colnames(bootstrapCoeffsBlack),
                                     names_to="met", values_to="coeff")

plotScaledEffects(bootstrapCoeffsWhite, bootstrapCoeffsBlack,
                  effect="main", type="LASSO", transformation="Log")


}
