% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmokingContributions.R
\name{plotSmokingContributions}
\alias{plotSmokingContributions}
\title{Visualize the main metabolite contributions in the discrepancies in prediction
between smokers and non-smokers.}
\usage{
plotSmokingContributions(data_smoking, data_test, model)
}
\arguments{
\item{data_smoking}{A data frame with metabolite measurements and predicted BMI of smoking patients}

\item{data_test}{A data frame with metabolite measurements and predicted BMI of non-smoking test set patients}
}
\value{
A ggplot object visualizing the main metabolic contributions of smoking to the discrepancy in predicted BMI
}
\description{
Visualize the main metabolite contributions in the discrepancies in prediction
between smokers and non-smokers.
}
