% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledEffects.R
\name{scaledEffects}
\alias{scaledEffects}
\title{Calculates regression coefficients of a model formulation and bootstrap replicates of the data set.
The predictors are scaled to unit variance in order to compare the regression coefficients of different predictors equally}
\usage{
scaledEffects(data, effect, type, transformation, balancing = "", boot.n = 100)
}
\arguments{
\item{data}{Data to draw bootstrap replicates from}

\item{effect}{"main" or "interaction" dependent on which effects are wished}

\item{type}{Regression type, must be "Ridge" or "LASSO for this function}

\item{transformation}{"Log" or "Inv", dependent on whether regression on log(BMI) or 1/BMI is required}

\item{balancing}{Defaults to "", not oversampling the data. Specify "Balanced" if oversampling is needed.}

\item{boot.n}{the amount of bootstrap replicates to be drawn}
}
\value{
A data tibble with regression coefficients for every replicate.
}
\description{
Calculates regression coefficients of a model formulation and bootstrap replicates of the data set.
The predictors are scaled to unit variance in order to compare the regression coefficients of different predictors equally
}
