% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableBiometrics.R
\name{tableBiometrics}
\alias{tableBiometrics}
\title{Make a table of patient counts for every ethnicity inside every obesity class}
\usage{
tableBiometrics(df)
}
\arguments{
\item{df}{A data frame containing columns "ObesityClass" and "Race".}
}
\value{
A matrix with all the observed ethnicities as columns, the different obesity classes and the ethnicity total as rows and the counts as entries.
}
\description{
Make a table of patient counts for every ethnicity inside every obesity class
}
\examples{

df_normal <- data.frame(ObesityClass=rep("Class1", 40), Race=c(rep("Race1", 15), rep("Race2", 25))
df_obese <- data.frame(ObesityClass=rep("Class2", 50), Race=c(rep("Race1", 20), rep("Race2", 30)))
df_all <- bind_rows(df_normal, df_obese)
tableBiometrics(df_all)

}
