% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSmokingStatus.R
\name{tableSmokingStatus}
\alias{tableSmokingStatus}
\title{Make a table of patient counts of smokers and non-smokers for every ethnicity}
\usage{
tableSmokingStatus(df)
}
\arguments{
\item{df}{A data frame containing columns "Smoking status" and "Race".}
}
\value{
A matrix with all the observed ethnicities as columns, counted "Smoking" and "Non-smoking" patients as entries.
}
\description{
Make a table of patient counts of smokers and non-smokers for every ethnicity
}
