% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulateValidation.R
\name{tabulateValidation}
\alias{tabulateValidation}
\title{Use 4-fold cross validation inside new.data to score the provided regression formulations
For every regression formulation, calculate the AUC of every fold building a model from the other 3 folds.
For every regression formulation, return the average AUC over the 4 folds.}
\usage{
tabulateValidation(effects, types, transformations, balancing = NULL, new.data)
}
\arguments{
\item{effects}{a vector of "main" and/or "interaction" effects of every formulation}

\item{types}{a vector of regression types "OLS", "Ridge" and/or "LASSO" of every formulation}

\item{balancing}{defaults to NULL, or a vector of "" and/or "Balanced" reflecting balancing requests of every formulation}

\item{new.data}{Data to build a regression model from}

\item{transformation}{a vector of BMI transformations "Log" and/or "Inv" of every formulation}
}
\value{
An averaged AUC value representing a score of the model formulation
}
\description{
Use 4-fold cross validation inside new.data to score the provided regression formulations
For every regression formulation, calculate the AUC of every fold building a model from the other 3 folds.
For every regression formulation, return the average AUC over the 4 folds.
}
\examples{
metabolites <- c("met1", "met2", "met_coll", "met_110")

formulations <- expand.grid(effects=c("main", "interaction"),
                            types=c("OLS", "Ridge", "LASSO"),
                            transformations=c("Log", "Inv"),
                            balancing=c("", "Balanced"),
                            stringsAsFactors=FALSE)

formulationScores <- tabulateValidation(formulations$effects, formulations$types,
                                        formulations$transformations,
                                        balancing=formulations$balancing, df)

View(formulationScores)

}
