% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testReclassificationIndependence.R
\name{testReclassificationIndependence}
\alias{testReclassificationIndependence}
\title{Test for reclassification independence between two tables of reclassification healthy/syndromatic counts by ethnicity.}
\usage{
testReclassificationIndependence(classCounts1, classCounts2)
}
\arguments{
\item{classCounts1}{A tibble or data frame containing columns "pred.: Healthy" and "pred.: Syndr."}

\item{classCounts2}{A second table with the same number of rows and columns "pred.: Healthy" and "pred.: Syndr."}
}
\value{
A chi-square p-value for the test of independence in reclassification between the two reclassification tables
}
\description{
Test for reclassification independence between two tables of reclassification healthy/syndromatic counts by ethnicity.
}
\examples{

library(dplyr)
library(tidyr)
library(tibble)

normals1 <- data.frame(ObesityClass=rep("Normal weight", times=50),
                      Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.2)])
overweights1 <- data.frame(ObesityClass=rep("Overweight", times=50),
                          Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.5)])
obeses1 <- data.frame(ObesityClass=rep("Obese", times=50),
                          Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.8)])
data1 <- bind_rows(normals1, overweights1, obeses1)
classCounts1 <- countClassesNumeric(data1, race="Not specified", strat="Unknown")

normals2 <- data.frame(ObesityClass=rep("Normal weight", times=50),
                      Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.1)])
overweights2 <- data.frame(ObesityClass=rep("Overweight", times=50),
                          Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.4)])
obeses2 <- data.frame(ObesityClass=rep("Obese", times=50),
                          Reclassified=c("Healthy", "Syndromatic")[1+rbinom(n=50, size=1, prob=0.8)])
data2 <- bind_rows(normals2, overweights2, obeses2)
classCounts2 <- countClassesNumeric(data2, race="Not specified", strat="Also Unknown")

testReclassificationIndependence(classCounts1, classCounts2)

}
