% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainRidgeLasso.R
\name{trainRidgeLASSO}
\alias{trainRidgeLASSO}
\title{Train a Ridge or LASSO model from the data with a specification about main/interaction effects and a BMI transformation.}
\usage{
trainRidgeLASSO(
  effect,
  type,
  transformation,
  new.data = NULL,
  new.x = NULL,
  new.y = NULL,
  lambda = NULL
)
}
\arguments{
\item{effect}{"main" or "interaction" dependent on which effects are wished}

\item{type}{Regression type, must be "Ridge" or "LASSO" for this function}

\item{transformation}{"Log" or "Inv", dependent on whether regression on log(BMI) or 1/BMI is required}

\item{new.data}{Data to build a regression model from}

\item{balancing}{Defaults to "", not oversampling the data. Specify "Balanced" if oversampling is needed.}
}
\value{
A glmnet regression model with 1 lambda parameter
}
\description{
Train a Ridge or LASSO model from the data with a specification about main/interaction effects and a BMI transformation.
}
\examples{
metabolites <- c("met1", "met2", "met_coll", "met_110")

data_train <- oversample(df)
trainRidgeLASSO(effect="main", type="Ridge", transformation="Log", new.data=data_train)

}
