% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateModelOLS.R
\name{validateModelOLS}
\alias{validateModelOLS}
\title{Use 4-fold cross validation inside new.data to score a regression formulation with type OLS.
Calculate the AUC of every fold building a model from the other 3 folds.
Return the average AUC over the 4 folds.}
\usage{
validateModelOLS(effect, type, transformation, balancing, new.data)
}
\arguments{
\item{effect}{"main" or "interaction" dependent on which effects are wished}

\item{type}{Regression type, must be "OLS" for this function}

\item{transformation}{"Log" or "Inv", dependent on whether regression on log(BMI) or 1/BMI is required}

\item{balancing}{Defaults to "", not oversampling the data. Specify "Balanced" if oversampling is needed.}

\item{new.data}{Data to build a regression model from}
}
\value{
An averaged AUC value representing a score of the model formulation
}
\description{
Use 4-fold cross validation inside new.data to score a regression formulation with type OLS.
Calculate the AUC of every fold building a model from the other 3 folds.
Return the average AUC over the 4 folds.
}
